dom/workers/test/test_ipAddressOrigin.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Web Workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="go();">
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<iframe id="testFrame"></iframe>
<script class="testbody" type="text/javascript">

  function go() {
    var iframe = document.getElementById("testFrame");

    iframe.addEventListener("load", function() {
      iframe.contentWindow.postMessage("go", "*");
      window.addEventListener("message", function(event) {
        is(event.data, "done", "Correct message");
        SimpleTest.finish();
      }, "false");
    }, false);

    iframe.src =
      "http://127.0.0.1" +
      window.location.pathname.replace("test_ipAddressOrigin.html",
                                       "ipAddressOrigin_iframe.html");
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
